+++ opentelemetry-sdk/src/metrics/exporter.rs
@@ -1,11 +1,116 @@
 //! Interfaces for exporting metrics
 
-use crate::error::OTelSdkResult;
-use std::time::Duration;
+use opentelemetry::InstrumentationScope;
 
-use crate::metrics::data::ResourceMetrics;
+use crate::{error::OTelSdkResult, Resource};
+use std::{fmt::Debug, slice::Iter, time::Duration};
 
-use super::Temporality;
+use super::{
+    data::AggregatedMetrics,
+    reader::{ResourceMetricsData, ScopeMetricsData},
+    InstrumentInfo, Temporality,
+};
+
+/// Stores borrowed metrics and provide a way to collect them
+#[derive(Debug)]
+pub struct ScopeMetricsCollector<'a> {
+    iter: ScopeMetricsLendingIter<'a>,
+}
+
+impl ScopeMetricsCollector<'_> {
+    /// Start collecting all metrics
+    pub fn collect(self, process: impl FnOnce(ScopeMetricsLendingIter<'_>)) {
+        process(self.iter)
+    }
+}
+
+/// A collection of [`ScopeMetricsCollector`] and the associated [Resource] that created them.
+#[derive(Debug)]
+pub struct ResourceMetrics<'a> {
+    /// The entity that collected the metrics.
+    pub resource: &'a Resource,
+    /// The collection of metrics with unique [InstrumentationScope]s.
+    pub scope_metrics: ScopeMetricsCollector<'a>,
+}
+
+/// Iterator over libraries instrumentation scopes ([`InstrumentationScope`]) together with metrics.
+/// Doesn't implement standard [`Iterator`], because it returns borrowed items. AKA "LendingIterator".
+pub struct ScopeMetricsLendingIter<'a> {
+    iter: Iter<'a, ScopeMetricsData>,
+}
+
+/// A collection of metrics produced by a [`InstrumentationScope`] meter.
+#[derive(Debug)]
+pub struct ScopeMetrics<'a> {
+    /// The [InstrumentationScope] that the meter was created with.
+    pub scope: &'a InstrumentationScope,
+    /// The list of aggregations created by the meter.
+    pub metrics: MetricsLendingIter<'a>,
+}
+
+/// Iterator over aggregations created by the meter.
+/// Doesn't implement standard [`Iterator`], because it returns borrowed items. AKA "LendingIterator".
+pub struct MetricsLendingIter<'a> {
+    iter: Iter<'a, super::reader::MetricsData>,
+}
+
+/// A collection of one or more aggregated time series from an [Instrument].
+///
+/// [Instrument]: crate::metrics::Instrument
+#[derive(Debug)]
+pub struct Metric<'a> {
+    /// The name of the instrument that created this data.
+    pub instrument: &'a InstrumentInfo,
+    /// The aggregated data from an instrument.
+    pub data: &'a AggregatedMetrics,
+}
+
+impl<'a> ResourceMetrics<'a> {
+    pub(crate) fn new(data: &'a ResourceMetricsData) -> Self {
+        Self {
+            resource: &data.resource,
+            scope_metrics: ScopeMetricsCollector {
+                iter: ScopeMetricsLendingIter {
+                    iter: data.scope_metrics.iter(),
+                },
+            },
+        }
+    }
+}
+
+impl Debug for ScopeMetricsLendingIter<'_> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("BatchScopeMetrics").finish()
+    }
+}
+
+impl ScopeMetricsLendingIter<'_> {
+    /// Advances the iterator and returns the next value.
+    pub fn next_scope_metrics(&mut self) -> Option<ScopeMetrics<'_>> {
+        self.iter.next().map(|item| ScopeMetrics {
+            scope: &item.scope,
+            metrics: MetricsLendingIter {
+                iter: item.metrics.iter(),
+            },
+        })
+    }
+}
+
+impl MetricsLendingIter<'_> {
+    /// Advances the iterator and returns the next value.
+    pub fn next_metric(&mut self) -> Option<Metric<'_>> {
+        self.iter.next().map(|item| Metric {
+            instrument: &item.instrument,
+            data: &item.data,
+        })
+    }
+}
+
+impl Debug for MetricsLendingIter<'_> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("BatchMetrics").finish()
+    }
+}
 
 /// Exporter handles the delivery of metric data to external receivers.
 ///
@@ -18,7 +123,7 @@ pub trait PushMetricExporter: Send + Sync + 'static {
     /// considered unrecoverable and will be logged.
     fn export(
         &self,
-        metrics: &ResourceMetrics,
+        metrics: ResourceMetrics<'_>,
     ) -> impl std::future::Future<Output = OTelSdkResult> + Send;
 
     /// Flushes any metric data held by an exporter.